requirements.txt
snowflake-connector-python>=3.0.0
cryptography>=41.0.0
